% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dune_palette.R
\name{dune_palette}
\alias{dune_palette}
\title{Creates color palettes inspired by "Dune"}
\usage{
dune_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette. Run \code{names(dune_palettes)} to view options.}

\item{n}{Number of desired colors: If the requested number of colors exceeds the available palette,
additional colors will be automatically interpolated. If the number of colors (n) is not specified,
the default length of the palette will be used.}

\item{type}{Specify either "continuous" or "discrete".
Use "continuous" if you want to automatically interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
Creates color palettes inspired by "Dune"
}
\examples{
dune_palette("arrakis")
}
