% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{dune_palettes}
\alias{dune_palettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 16.
}
\usage{
dune_palettes
}
\description{
Use names(dune_palettes) to return all palette names.
}
\keyword{datasets}
