% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdune-package.R
\docType{package}
\name{Rdune-package}
\alias{Rdune}
\alias{Rdune-package}
\title{Rdune: 'Creates Color Palettes Inspired by Dune'}
\description{
Enables the use of color palettes inspired by the 'Dune' movies. These palettes are compatible with 'ggplot2'. See Wickham (2016) \doi{10.1007/978-3-319-24277-4} for more details on 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nvietto/Rdune}
  \item \url{https://nvietto.github.io/Rdune/}
  \item Report bugs at \url{https://github.com/nvietto/Rdune/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Vietto \email{nicholasvietto@gmail.com} [copyright holder]

}
\keyword{internal}
