% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_dune_c.R
\name{scale_color_dune_c}
\alias{scale_color_dune_c}
\title{Dune palettes for plotting with 'ggplot2'}
\usage{
scale_color_dune_c(name, ...)
}
\arguments{
\item{name}{Name of palette. Run \code{names(dune_palettes)} to view options}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_color_gradientn}}}
}
\value{
An object defining a continuous color scale for use with ggplot2.
}
\description{
Dune palettes for plotting with 'ggplot2'
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = mpg, y = disp, color = hp)) +
geom_point() +
scale_color_dune_c("atreides3")
}
