% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_dune_c.R
\name{scale_fill_dune_c}
\alias{scale_fill_dune_c}
\title{Dune palettes for plotting with ggplot2}
\usage{
scale_fill_dune_c(name, ...)
}
\arguments{
\item{name}{Name of palette. Run \code{names(dune_palettes)} to view options}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_fill_gradientn}}}
}
\value{
An object defining a continuous fill scale for use with ggplot2.
}
\description{
Dune palettes for plotting with ggplot2
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = mpg, y = disp, fill = wt)) +
geom_point() +
scale_fill_dune_c("sardaukar")
}
