% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dune_palettes.R
\docType{data}
\name{dune_palettes}
\alias{dune_palettes}
\title{Complete list of palettes.
Use names(dune_palettes) to return all palette names.}
\format{
An object of class \code{list} of length 18.
}
\usage{
dune_palettes
}
\description{
Complete list of palettes.
Use names(dune_palettes) to return all palette names.
}
\keyword{datasets}
